refactor outbox events
