+++ libs/core/src/integration/index.ts
@@ -1,3 +1,2 @@
 export * from './email.schemas';
 export * from './notifications.schemas';
-export * from './outbox.schema';
