+++ libs/schemas/src/events/index.ts
@@ -1,14 +1,36 @@
 import { z } from 'zod';
+import { PG_INT } from '../utils';
 import { events } from './events.schemas';
 
 export type EventSchemas = typeof events;
 export type Events = keyof EventSchemas;
-export type EventPair<K extends Events> = {
+
+// TODO: clean list of events that should not be emitted to the outbox
+export type OutboxSchemas = Omit<EventSchemas, 'ContestRolloverTimerTicked'>;
+export type OutboxEvents = keyof OutboxSchemas;
+export type EventPair<K extends OutboxEvents> = {
   event_name: K;
-  event_payload: z.infer<EventSchemas[K]>;
+  event_payload: z.infer<OutboxSchemas[K]>;
 };
 export type EventPairs = {
-  [K in Events]: EventPair<K>;
-}[Events];
+  [K in OutboxEvents]: EventPair<K>;
+}[OutboxEvents];
+
+export const outboxEvents = Object.keys(events) as OutboxEvents[];
+export const Outbox = z.union(
+  outboxEvents.map((event_name) =>
+    z
+      .object({
+        event_name: z.literal(event_name),
+        event_payload: events[event_name],
+      })
+      .extend({
+        event_id: PG_INT.optional(),
+        relayed: z.boolean().optional(),
+        created_at: z.coerce.date().optional(),
+        updated_at: z.coerce.date().optional(),
+      }),
+  ) as unknown as readonly [z.ZodTypeAny, z.ZodTypeAny, ...z.ZodTypeAny[]],
+);
 
 export { events };
