+++ packages/commonwealth/server/bindings/relay.ts
@@ -1,5 +1,6 @@
-import { Broker, Outbox, logger, stats } from '@hicommonwealth/core';
+import { Broker, logger, stats } from '@hicommonwealth/core';
 import type { DB } from '@hicommonwealth/model';
+import { Outbox } from '@hicommonwealth/schemas';
 import { QueryTypes } from 'sequelize';
 import { z } from 'zod';
 import { config } from '../config';
